requirements.txt
Adafruit-Blinka==8.47.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-motor==3.4.13
adafruit-circuitpython-pca9685==3.4.15
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-servokit==1.3.17
adafruit-circuitpython-typing==1.11.0
Adafruit-PlatformDetect==3.73.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
